Move legacy tokenizers to `tokenizers.legacy`
