+++ curated_transformers/tests/tokenization/legacy/test_bert_tokenizer.py
@@ -2,17 +2,20 @@ import pytest
 import torch
 
 from curated_transformers._compat import has_hf_transformers
-from curated_transformers.tokenizers import BertTokenizer, PiecesWithIds
-from curated_transformers.tokenizers.bert_tokenizer import BertPreEncoder
+from curated_transformers.tokenizers import PiecesWithIds
 from curated_transformers.tokenizers.chunks import (
     InputChunks,
     SpecialPieceChunk,
     TextChunk,
 )
+from curated_transformers.tokenizers.legacy.bert_tokenizer import (
+    BertPreEncoder,
+    BertTokenizer,
+)
 from curated_transformers.tokenizers.tokenizer import DefaultNormalizer
 
-from ..util import torch_assertclose
-from .util import compare_tokenizer_outputs_with_hf_tokenizer
+from ...util import torch_assertclose
+from ..util import compare_tokenizer_outputs_with_hf_tokenizer
 
 
 @pytest.fixture
