+++ curated_transformers/tests/tokenization/legacy/test_camembert_tokenizer.py
@@ -3,10 +3,12 @@ import torch
 
 from curated_transformers._compat import has_hf_transformers
 from curated_transformers.tokenizers import PiecesWithIds
-from curated_transformers.tokenizers.camembert_tokenizer import CamembertTokenizer
+from curated_transformers.tokenizers.legacy.camembert_tokenizer import (
+    CamembertTokenizer,
+)
 
-from ..util import torch_assertclose
-from .util import compare_tokenizer_outputs_with_hf_tokenizer
+from ...util import torch_assertclose
+from ..util import compare_tokenizer_outputs_with_hf_tokenizer
 
 
 @pytest.fixture
