+++ curated_transformers/tokenizers/legacy/camembert_tokenizer.py
@@ -5,9 +5,9 @@ from curated_tokenizers import SentencePieceProcessor
 
 from curated_transformers.tokenizers.hf_hub import LegacyFromHFHub
 
+from ..tokenizer import AddBosEosPreEncoder
 from ._fairseq import FAIRSEQ_PIECE_IDS, FairSeqPostEncoder, FairSeqPreDecoder
 from .sentencepiece_tokenizer import SentencePieceTokenizer
-from .tokenizer import AddBosEosPreEncoder
 
 # Only provided as typing.Self in Python 3.11+.
 Self = TypeVar("Self", bound="CamembertTokenizer")
