+++ curated_transformers/tokenizers/legacy/roberta_tokenizer.py
@@ -1,15 +1,14 @@
-import json
 from pathlib import Path
-from typing import Any, Dict, Iterable, List, Optional, Tuple, Type, TypeVar, cast
+from typing import Any, Dict, Iterable, List, Optional, Tuple, Type, TypeVar
 
 from curated_tokenizers import ByteBPEProcessor
 
 from curated_transformers.tokenizers.hf_hub import LegacyFromHFHub
 
+from ..hf_hub import LegacyFromHFHub
+from ..tokenizer import AddBosEosPreEncoder, PreDecoder
+from ..util import remove_pieces_from_sequence
 from .bbpe_tokenizer import ByteBPETokenizer
-from .hf_hub import LegacyFromHFHub
-from .tokenizer import AddBosEosPreEncoder, PreDecoder
-from .util import remove_pieces_from_sequence
 
 # Only provided as typing.Self in Python 3.11+.
 Self = TypeVar("Self", bound="RobertaTokenizer")
