+++ curated_transformers/tokenizers/legacy/xlmr_tokenizer.py
@@ -3,10 +3,10 @@ from typing import Any, Dict, Optional, Type, TypeVar
 
 from curated_tokenizers import SentencePieceProcessor
 
+from ..hf_hub import LegacyFromHFHub
+from ..tokenizer import AddBosEosPreEncoder
 from ._fairseq import FAIRSEQ_PIECE_IDS, FairSeqPostEncoder, FairSeqPreDecoder
-from .hf_hub import LegacyFromHFHub
 from .sentencepiece_tokenizer import SentencePieceTokenizer
-from .tokenizer import AddBosEosPreEncoder
 
 # Only provided as typing.Self in Python 3.11+.
 Self = TypeVar("Self", bound="XlmrTokenizer")
